# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Claude Code Zen API
  description: |
    Comprehensive API documentation for Claude Code Zen - AI-powered development coordination platform.
    
    This API provides endpoints for:
    - System health and status monitoring
    - Swarm management and coordination
    - Task management and orchestration
    - Document and file operations
    - LLM statistics and analytics
    - Matron advisory system
    - Visionary roadmap planning
    - Workspace management
    - Real-time WebSocket communications
    
    The API follows RESTful principles and returns JSON responses.
  version: 2.0.0
  contact:
    name: Claude Code Zen Team
    url: https://github.com/zen-neural/claude-code-zen
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://fra-d1.in.centralcloud.net/api
    description: Production server
paths:
...
  /status:
    get:
      summary: System Status
      description: Get comprehensive system status including performance metrics
      operationId: getSystemStatus
      tags:
        - System
      responses:
        '200':
          description: System status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStatus'
        '500':
          description: Failed to get system status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Swarm Management
...
  /workspace/files:
    get:
      summary: List Files
      description: List files and directories in the workspace
      operationId: listWorkspaceFiles
      tags:
        - Workspace
      parameters:
        - name: path
          in: query
          description: Directory path to list (relative to workspace root)
          required: false
          schema:
            type: string
            default: ""
      responses:
        '200':
          description: Files listed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to list files
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create File
      description: Create a new file in the workspace
      operationId: createWorkspaceFile
      tags:
        - Workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFileRequest'
            example:
              path: "src/components/NewComponent.tsx"
              content: "import React from 'react';\n\nexport const NewComponent = () => {\n  return <div>Hello World</div>;\n};"
              template: "react"
      responses:
        '201':
          description: File created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileOperationResponse'
        '400':
          description: Invalid file data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to create file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update File
      description: Update content of an existing file
      operationId: updateWorkspaceFile
      tags:
        - Workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFileRequest'
            example:
              path: "src/components/Component.tsx"
              content: "// Updated file content\nimport React from 'react';\n\nexport const Component = () => {\n  return <div>Updated!</div>;\n};"
      responses:
        '200':
          description: File updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileOperationResponse'
        '400':
          description: Invalid file data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to update file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Delete File
      description: Delete a file from the workspace
      operationId: deleteWorkspaceFile
      tags:
        - Workspace
      parameters:
        - name: path
          in: query
          description: File path to delete (relative to workspace root)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileOperationResponse'
        '400':
          description: File path required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to delete file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Resource not found"
        code:
          type: string
          description: Error code (optional)
          example: "NOT_FOUND"
        details:
          type: object
          description: Additional error details (optional)

    # System Status
    SystemStatus:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
          example: "healthy"
        timestamp:
          type: string
          format: date-time
        uptime:
          type: number
        performance:
          $ref: '#/components/schemas/PerformanceMetrics'
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceStatus'
    PerformanceMetrics:
      type: object
      properties:
        cpu:
          type: number
          format: float
          description: CPU usage percentage
          example: 15.5
        memory:
          type: number
          format: float
          description: Memory usage percentage
          example: 45.2
        disk:
          type: number
          format: float
          description: Disk usage percentage
          example: 60.1
        responseTime:
          type: number
          description: Average response time in milliseconds
          example: 125
    ServiceStatus:
      type: object
      properties:
        name:
          type: string
          example: "swarm-coordinator"
        status:
          type: string
          enum: [running, stopped, error]
          example: "running"
        lastCheck:
          type: string
          format: date-time
        details:
          type: object

    # Swarm Management
...
    FileInfo:
      type: object
      properties:
        name:
          type: string
          example: "component.tsx"
        path:
          type: string
          example: "src/components/component.tsx"
        type:
          type: string
          enum: [file, directory]
          example: "file"
        size:
          type: integer
          description: File size in bytes
          example: 2048
        modified:
          type: string
          format: date-time
        extension:
          type: string
          example: "tsx"
        isHidden:
          type: boolean
          example: false
    FileListResponse:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/FileInfo'
        currentPath:
          type: string
          example: "src/components"
...
    CreateFileRequest:
      type: object
      required:
        - path
      properties:
        path:
          type: string
          example: "src/components/NewComponent.tsx"
        content:
          type: string
          default: ""
          example: "import React from 'react';\n\nexport const NewComponent = () => {\n  return <div>New Component</div>;\n};"
        template:
          type: string
          enum: [typescript, react, svelte, markdown]
          example: "react"
    UpdateFileRequest:
      type: object
      required:
        - path
        - content
      properties:
        path:
          type: string
          example: "src/components/component.tsx"
        content:
          type: string
          example: "// Updated content\nimport React from 'react';\n\nexport const Component = () => {\n  return <div>Updated!</div>;\n};"
...
    FileOperationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        path:
          type: string
          example: "src/components/component.tsx"
        message:
          type: string
          example: "File created successfully"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (if implemented)

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication (if implemented)
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...